move dimensions to core
