+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-arbitrary-utilities.ts
@@ -3,8 +3,8 @@ import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
 import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
 import type { Writable } from '../../../../tailwindcss/src/types'
 import { DefaultMap } from '../../../../tailwindcss/src/utils/default-map'
+import { dimensions } from '../../../../tailwindcss/src/utils/dimensions'
 import * as ValueParser from '../../../../tailwindcss/src/value-parser'
-import { dimensions } from '../../utils/dimension'
 import { baseCandidate, parseCandidate } from './candidates'
 import { computeUtilitySignature, preComputedUtilities } from './signatures'
